api_myzeancoach/requirements.txt
six==1.8.0
oauthlib==0.7.2
Django==1.8.4
Markdown==2.5.2
django-uuidfield==0.5.0
django-oauth-toolkit==0.10.0
django-braces==1.4.0
django-filter==0.13
djangorestframework==3.2.2
drf-nested-routers==0.11.1
django-push-notifications==1.4.1
django-ordered-model==1.2.1
django-permissions==1.2.2
vatnumber==1.2
django-countries==4.0
pytz==2016.4  # required for Babel
snowballstemmer==1.2.1
imagesize==0.7.1  # required for Sphinx
docutils==0.12  # required for Sphinx
MarkupSafe==0.23  #required for Jinja2
Jinja2==2.8  # required for Sphinx
Pygments==2.1.3  # required for Sphinx
alabaster==0.7.8  # theme, required for Sphinx
Babel==2.3.4  # required for Sphinx
Sphinx==1.4.1
pockets==0.3  # required for sphinx-contrib-napoleon
sphinxcontrib-napoleon==0.5.0
click==6.6  # required by sphinx-intl
sphinx-intl==0.9.9
pycrypto==2.6.1
dj-database-url==0.3.0
django-cors-headers==1.1.0
requests==2.11.1
pydns==2.3.6
beautifulsoup4==4.5.3